Move `core` and `plugins` to `components` folder
