+++ components/app/app.go
@@ -6,21 +6,21 @@ import (
 	"github.com/iotaledger/hive.go/app"
 	"github.com/iotaledger/hive.go/app/components/profiling"
 	"github.com/iotaledger/hive.go/app/components/shutdown"
-	"github.com/iotaledger/wasp/core/chains"
-	"github.com/iotaledger/wasp/core/database"
-	"github.com/iotaledger/wasp/core/dkg"
-	"github.com/iotaledger/wasp/core/logger"
-	"github.com/iotaledger/wasp/core/nodeconn"
-	"github.com/iotaledger/wasp/core/peering"
-	"github.com/iotaledger/wasp/core/processors"
-	"github.com/iotaledger/wasp/core/publisher"
-	"github.com/iotaledger/wasp/core/registry"
-	"github.com/iotaledger/wasp/core/users"
-	"github.com/iotaledger/wasp/core/wasmtimevm"
+	"github.com/iotaledger/wasp/components/chains"
+	"github.com/iotaledger/wasp/components/database"
+	"github.com/iotaledger/wasp/components/dkg"
+	"github.com/iotaledger/wasp/components/logger"
+	"github.com/iotaledger/wasp/components/nodeconn"
+	"github.com/iotaledger/wasp/components/peering"
+	"github.com/iotaledger/wasp/components/processors"
+	"github.com/iotaledger/wasp/components/profilingrecorder"
+	"github.com/iotaledger/wasp/components/prometheus"
+	"github.com/iotaledger/wasp/components/publisher"
+	"github.com/iotaledger/wasp/components/registry"
+	"github.com/iotaledger/wasp/components/users"
+	"github.com/iotaledger/wasp/components/wasmtimevm"
+	"github.com/iotaledger/wasp/components/webapi"
 	"github.com/iotaledger/wasp/packages/toolset"
-	"github.com/iotaledger/wasp/plugins/profilingrecorder"
-	"github.com/iotaledger/wasp/plugins/prometheus"
-	"github.com/iotaledger/wasp/plugins/webapi"
 )
 
 var (
