+++ components/dkg/component.go
