+++ components/logger/component.go
