+++ components/publisher/component.go
