+++ components/registry/component.go
