+++ components/registry/params.go
