+++ tools/evm-server/waypoint.hcl
@@ -38,7 +38,7 @@ app "wasp-evm-server" {
             context    = "../.."
             build_args = {
                 GOLANG_IMAGE_TAG = "1.20-bullseye"
-                BUILD_LD_FLAGS = "-X=github.com/iotaledger/wasp/core/app.Version=${gitreftag()}"
+                BUILD_LD_FLAGS = "-X=github.com/iotaledger/wasp/components/app.Version=${gitreftag()}"
                 BUILD_TARGET = "./tools/wasp-cli"
                 FINAL_BINARY = "wasp-cli"
             }
