+++ tools/local-setup/build_container.cmd
@@ -1,6 +1,6 @@
 rem export variables for the build process
 set BUILD_TAGS=rocksdb
-for /f %%f in ('git describe --tags') do set BUILD_LD_FLAGS=-X=github.com/iotaledger/wasp/core/app.Version=%%f
+for /f %%f in ('git describe --tags') do set BUILD_LD_FLAGS=-X=github.com/iotaledger/wasp/components/app.Version=%%f
 
 rem build the wasp container
 docker compose build wasp
