+++ tools/local-setup/build_container.sh
@@ -5,7 +5,7 @@ GIT_REF_TAG="$(git describe --tags)"
 
 # export variables for the build process
 export BUILD_TAGS=rocksdb
-export BUILD_LD_FLAGS="-X=github.com/iotaledger/wasp/core/app.Version=${GIT_REF_TAG}"
+export BUILD_LD_FLAGS="-X=github.com/iotaledger/wasp/components/app.Version=${GIT_REF_TAG}"
 
 # build the wasp container
 docker compose build wasp
