+++ tools/local-setup/docker-compose.yml
@@ -78,7 +78,7 @@ services:
       dockerfile: Dockerfile.noncached
       args:
         BUILD_TAGS: "${BUILD_TAGS:-rocksdb}"
-        BUILD_LD_FLAGS: "${BUILD_LD_FLAGS:--X=github.com/iotaledger/wasp/core/app.Version=v0.0.0-testing}"
+        BUILD_LD_FLAGS: "${BUILD_LD_FLAGS:--X=github.com/iotaledger/wasp/components/app.Version=v0.0.0-testing}"
     ulimits:
       nofile:
         soft: 16384
